VIRO-2589 Added VROTrackingHelper to the VROARSessioniOS
This allows us to perform image recognition while using ARKit. However,
to remove the compiled dependency on OpenCV, the code is commented out
in VROARSessioniOS and two previous OpenCV related commits need to be
reverted before this code will work.

Once that's done, you should see that recognition is running in the
logs (no rendering).