VIRO-2628: Enable ViroSceneView Transparent Background
- Implemented a GLTextureView to be used by ViroViewScene to render the 3D scene,
instead of a GLSurfaceView. This is needed as experiences with Android views behind
and infront of a GLSurfaceView is not technically feasible, due to the limitations
of Android GLSurface views (it is actually a dedicated Surface Window). GLTextureView
mitigates this, with a slight performance cost that is not noticeable at this time,
while also enabling the ability to have a transparent background experience.

- Added a setBackgroundColor API within the ViroViewScene control. This enables
developers to determine the background of the renderer that contains Viro Controls,
and thus to set a fully opaque or transparent background (that exposes android views
laid out beneath the renderer). Internally, this would toggle the clearColor of the
appropriate render targets that are used in the main render pass.

- Modified shader programs to include an alpha channel when rendering the scene, so
that they avoid overwritting the transparent scene unnecessarily with opaquness.