VIRO-39 [PBR]: fix specular term
The specular term wasn't working because we had
divide by zero protection in the distributionGGX
function that cut off the highlights.

Also we forgot to multiply by radiance but that was just
an issue with the previous commit.

The balls still appear more "orange" that they should
but it's much closer to the reference now.