VIRO-2393: irradiance convolution
Generate the diffuse irradiance map by convolution
of the lighting environment. This involved creating a
new type of render pass, very similar to the equi-
rectangular to cube render pass. Abstracted out some
similarities into VRORenderUtil to make these kinds
of passes easier to build.

Temporarily, the diffuse irradiance is rendered as the
background of the scene.