VIRO-2805: use luminous flux for intensity
This commit changes the intensity parameter for
non-ambient lights to use luminous flux. Note we're
not *precisely* using luminous flux. Normally, the
way this works is you compute luminous intensity
(intensity over solid angle) from luminous flux (basically,
the power of the bulb). The problem is this makes
the intensity of the light dependent on the area of
emission of the light: for example, a spotlight would
be more intense the smaller its cone is, and an Omni
light, be extension, would be less intense than a similar
spotlight.

To make this easier for artists, but still remaining
somewhat physical, we instead simplify the formula
so that luminous intensity is just luminous flux divided
by PI. This is the technique used by Frostbite for spot
lights, and (apparently) by SceneKit for both spot and
point lights. We use it for both spot and point lights.

In the future when we introduce area lights, we may
go back to a more purely physical system.

Further changes were required to make this work:

1. Lighting modifiers no longer take a "shader light".
    That's been removed, now the lighting modifier
    itself computes the attenuation, since attenuation
    now depends on the lighting model.

2. Basic lighting models now compute attenuation
    themselves and multiply by intensity / 1000.0.

3. Physical lights now compute attenuation using
    physical properties (inverse square fallout), *ignoring*
    the attenuation start distance, since that's very much
    not physical. The end distance is still used both in
    culling and rendering.

4. Simplified the spotlight attenuation function so that it
    takes cosine values. Furthermore, realized our spotlight
    angles are misleading: they should follow the convention
    where outer_angle is the angle after which the light has
    zero impact. Previous this angle was inner_angle +
    outer_angle. Also, spotlight angles should be specified as
    the full cone angle, not half the cone angle.

5. Reformatted the PBR equations in the VROShaderFactory
    to reflect my better understanding of how it's all working.
    Also added a crucial "multiply by PI" at the end, which
    is required for point lights. Use proper photometric variable
    names instead of radiometric.

6. Updated use of shadow maps to reflect the new inner and
    outer angles used by spotlights.

7. Added new Photometric light test.