VIRO-2805: model directional light as non-physical
Directional light shouldn't try to use lumens as it's
far removed from anything physical. Instead if a
directional light is used, model the intensity as simply
a scalar divided by 1,000 (as we do for non-PBR
surfaces).

In the future we can model directional light as a
faraway disc area light, as per Frostbite.