VIRO-2831: customizable field of view
Refactor the field of view system so that:

+ FOV can be set in NodeCamera, so that users can set
   specific field of views for different cameras.

+ The NodeCamera FOV is used for ViroSceneView and
    other monocular views, but ignored in VR or AR mode,
    where instead the platform FOV is used.

+ If no FOV is specified, uses a *new default* FOV of
   60 degrees.

+ Introduce new mechanism of setting the FOV: you set the
   FOV across the "major axis" (e.g. the larger axis)
   instead of on X or Y. This makes the FOV setting robust
   to orientation changes.

Tested by grabbing the FOV from an ARCore view and
setting it in a SceneView, and verified they look identical.

Note the new default FOV is much smaller than the old one.
The one one was a large fisheye that caused distortion on the
edges. We can tweak this new default FOV prior to release
if it's gone too far in the other (narrow) direction.