VIRO-2783: Modify OBJ / FBX Materials from ViroCore Java
- Constructed a list of jMaterials to be passed back along with the
onObjLoad callback in java and then cached in Object3D.java. This
would enable developers to then modify the material properties of
a given FBX / OBJ model.
- Added jMaterial and jTexture contruction helper methods to create
the java objects for corresponding native VROMaterial and VROTexture
objects.
- Added jMaterial verification to the Object3D virocore tests.
- Fixed the Object3D tests as they were originally blocking on the
UI thread and causing the test to hang. Tests are still being run
on the render thread (not ideal), but at least is unblocked for now.