VIRO-2894: expose project and unproject methods
Expose methods projectPoint and unprojectPoint
in ViroView and VRORenderer. These methods
convert screen coordinates to world coordinate and
back. Tested to ensure that the methods correctly
undo one another; e.g. project(unproject(P)) = P.

The interesting part about this implementation is that
the standard glu functions do *not* undo one another
because the depth they return is exponential. It needs
to be linearized in order to work correctly. Our method
matches what's used in SceneKit and similar rendering
engines.