VIRO-2885: FBX stretching on Android devices
This was occurring on Android devices because we
were limiting our number of bones to 64. We were doing
this based on a commit on 10/2/17, where we noticed
crashes were occurring on Samsung/Adreno when
running skeletal animations, *unless* the bone UBO
size was reduced to 64. Back then we thought the
problem was in the Adreno driver.

However, later on we fixed a different bug, VIRO-2286,
where skeletal animation was crashing on OVR
unless the bone UBO was initialized to zero-ed out
data. It turns out that also fixes the above error, so we
can now move the bone UBO size back up to 192.

This also means we can close the TBO story (VIRO-1935)
since UBOs are working fine for skeletal animation.