VIRO-2834: Specular Looks too shiny with high reflectance
Turns out, the Hammersley sequence calculation in the prefilter shaders for IBL
fails on some devices without bit operator support, resulting in an incorrect
mipmap blur. This would result in a more-so reflective ball, even for rough
surfaces. The fix is to apply the no bit ops version of the Hammersley sequence.