VIRO-2921: JNI memory leaks + crashes
These were caused because jni.hpp holds
onto a JNI env, which causes instability when
we delete a local, global, or weak reference
on a different thread than that on which the object
was created. This occurs most often when
running the MemoryLeakTest, where we GC and
clean up memory.

Fix this by making jni.hpp use VROPlatformGetEnv
so it always gets a valid JNI Env, even if on the
finalizer thread.

Also replace release() with reset() on various JNI
objects. The reason we didn't see this bug before
is that we were actually leaking those objects because
release() is the wrong method to call on a unique_ptr.