VIRO-2942, VIRO-2943: Large OBJ Crashes
- Deleted unused JNI local refs for Texture_JNI.
- Loaded OBJ jMaterials and cleaned them up, one at a time, rather than
creating a list of jMaterials to be put set in an ObjectArrayElement. The
jMaterials are verified to be alive, even though we delete it's local ref
after setting it on the ObjectArrayElement. Thus, the array element in
JNI itself must be holding onto the ref, so we should be okay. In turn,
this should significantly lower the amount of local refs in JNI.
- Removed unused JNI local refs when setting jEnum values.
- Have elementIndex be of type uint32_t instead of unint8_t. Using an 8 bit
element index with large OBJs would cause a clobbering of material refs that
would result in an incorrectly bounded material element and ultimately a
crash.