VIRO-2464: workaround to GVR breaking recordings
GVR fixed the bug reference in VIRO-2464, but
continues to attempt to deactivate the app-wide
audio session periodically. This workaround makes
is so that GVR is lazily initialized, so that at least
apps that aren't using spatial sound can still
make audio recordings. New bug has filed on
Google to fix the new issue.

This commit also:

+ Updates GVR to the latest version, so that
   the referenced JIRA issue is fixed.
+ Makes VROSoundGVR use a weak pointer
   to the GVR audio context. This is a better
   pattern in general so we can (if desired) turn
   off GVR entirely after a recording starts. We
   don't do this yet but it's a possibility for a
   better workaround.