VIRO-2946: ARCameraHitTest Bug
The current VROProjector projects / un-projects a 3D Vec point, however for the
arCameraHit test, a ray is provided to be shot from the position of the camera (bug).
Because Arkit needs a 2D screen point, we'll need to project the provided ray at the
current camera's position to be then passed on to the ArCore's hit test.