VIRO-2957: migrate to ARCore native API
To fix performance issues on older devices, we
have migrated to the native ARCore API. This
greatly reduces the stress on the garbage collector,
which was slowing down performance on the
Samsung A7. Google's ARCore Java API uses too
many rapidly created objects with finalizers, which
stalls the GC and results in OOM crashes. The
problem disappears with this new native API usage.

All functionality appears to work but further testing
is needed. Temporarily disabled the AR supported
checks.

Also of note is the ARCore AAR file has been edited,
the .so files are removed and placed in the jniLibs
folder.