VIRO-2965, VIRO-2958: typeface fixes
The introduction of typeface caching exposed bugs on
iOS because we were releasing the data associated
with the font, and Freetype requires that data remain in
memory until the tyepface is disposed of.

Also started using a weak cache for the typefaces so that
they are disposed when no longer used. This keeps the
total number of typefaces (and glyphs) to a minimum,
improving memory consumption.