Fix a bug with the calculation of bounding boxes
Bounding boxes are in world coordinates, so they should never be
initialized to 0,0,0,0,0,0 because it would mask any negative numbers
in a max(a,b) scenario and vice versa for positive numbers. Change the
renderer to always update the bounding boxes with the node's position.

This issue was causing us to get _umbrellaBoundingBoxes with values like:

-1.25, 0, -1.25, 0, -1.25, 0

even though the object is positioned at [-1, -1, -1], with the fix in, the
bounding box is now:

-1.25, -.75, -1.25, -.75, -1.25, -.75