Added setAtomicRenderProperties() func to VRONode
This function should be called when the render properties have
all settled (ie. at the end of the compute phase of the render pass).
Previously this was causing issues where some of the atomic properties
were being set 1 frame after they were computed because they were being
set in applyConstraints().