VIRO-2982: separate viro_arcore into separate lib
Create a new shared library, viro_arcore, that depends
on viro_native. The primary work involved here was
adding inheritance to Renderer.java, so that
RendererARCore.java (and its native counterpart) can
contain all the ARCore specific functionality.

Unfortunately, while the code to actually separate ARCore
out from Viro native was easy (because we'd already
compartmentalized ARCore from the renderer early-on
when we created generalized AR interfaces), the process
of exporting *two* shared libraries caused a host of
problems.

1. We can no longer use static STL when we're exporting
two shared libraries. Doing so creates random heap corruption
as the two STLs conflict. As a result, we had to move to using
shared STL -- hence the change in build.gradle.

2. Ug shared STL means using libgnustl_shared.so. Sadly,
React Native *also* uses GNU STL, but an older version: 4.8.
We use 4.9. This causes a conflict that makes react-viro fail
to start.

This commit does not resolve the second issue, leaving
react-viro broken. The long-term solution is going to be
recompiling all of our code to use GCC 4.8, or somehow
recompiling React Native to use 4.9, or perhaps using the
more recent libc++ build if it's stable. I did manage to recompile
protobufs to use libc++, but running ViroCore with libc++
caused strange JNI crashes.

The next commit will come with the short-term solution to
this problem, which is a separate build-file for react-viro.
Essentially, only ViroCore will export two shared libraries;
for viro-react, we'll continue to export a single library and use
static STL. This is ok because only ViroCore (for now) has a
SceneView. In ViroReact, we're for the most part AR-only
so we don't have to worry about older devices, and hence
don't have to worry about splitting ARCore into a separate
lib.