VIRO-2982: unified lib for viro-react
To solve the problem described in the last commit,
where ViroCore requires separate libs for SceneView
and ViroViewARCore while ViroReact requires a
single lib in order to use static STL, this commit does
the following:

1. Adds a new native library, viro_react, which is a union
of viro_native and viro_arcore.

2. Makes the viro_core flavor build viro_native and
viro_arcore.

3. Makes the viro_react flavor build viro_react.

4. Change all views so that they first attempt to load
viro_native and, if that fails, load viro_react.

This way the two AAR files (viro_core and viro_react)
have the library scheme that works for them.