VIRO-2833: Render Fixed Point cloud
- Created a VROFixedParticleEmitter class that uses the particle system
to draw a group of fixed points, in the form of bill-boarded quads, given
a vector of particle positions in world space.

- Refactored VROPointCloudEmitter into a generic VROFixedParticleEmitter
class and removed all point cloud and AR logic implementation into the
ARScene instead. An ARScene will contain and configure a fixed particle
emitter for rendering it's own point clouds.

- Plumbed VROFixedParticleEmitter through JNI for ViroCore.