VIRO-2355 [wasm]: make non-PBR shaders work
In WebGL, shaders can mysteriously "not work" which
causes objects to not appear. The only way to debug
this is to iteratively strip down the shader until the
objects start to appear again. This is how we discovered
that a few of our constructs do not work with WebGL:

1. The particles fragment shader include
2. The skinning vertex shader include
3. Our assigning of bone indices to the bone structure

Number 3 makes sense as that was always a little
iffy (assigning a non-existent attribute to a struct). The
others probably have to do with the large amounts of
memory they take up in lengthy arrays.

PBR shaders have not yet been tested.