VIRO-3075: make resource loading async
This is a large refactor that seeks to make all resource
loading asynchronous. Previously, when loading OBJ
and FBX files, we would load resources on a background
thread, using blocking calls on that background thread.
This was acceptable on those platforms since they
supported full multithreading -- it was ok to just block a
background thread while downloading resources required
for an OBJ, for example.

However, with WebAssembly we don't have that kind
of multithreading; instead, we can only do downloads,
etc on a background thread. As a result, we require a
system where we can do things asynchronously where
available, and fallback to single-threaded mode where
not available.

This commit achieves this by introducing a new
VROTaskQueue class, which can be used to queue up
tasks to be run serially or concurrently. When all tasks
are complete, it signals the rendering thread. We use this
new class to break down the FBX and OBJ loaders into
discrete tasks, each triggered from the rendering thread
but run in background threads. When these tasks are
all complete, we successfully load the model.

This introduced several other difficulties:

1. The OBJ loader had to be rewritten so that internally
    it can wait for MTL files to be downloaded and read.
    While doing this I simplified the OBJ loader considerably,
    deleting un-used methods (there were many) and
    reformatting the class to 4-space tabs.

2. The former "processResource" method in VROModelIOUtil
    was replace by "retrieveResourceAsync", which
    asynchronously grabs any resource, storing it in a file.

3. Concurrent execution of background tasks revealed a bug
    where do not give temporary files created out of Android
    assets unique names. Fixed this by adding a
    RandomString class to generate unique name suffixes.

4. Removed the "async" flag on the OBJ and FBX loaders,
    since with the new system we can always use the renderer
    callbacks to achieve similar results.

5. Reworked several callbacks so that they always dump
    us back on the rendering thread. Using this paradigm of
    always calling back to the rendering thread *greatly*
    simplifies the API use. This enabled the task queue to work
    without any atomic or blocking operations (as it can
    leverage those which already exist on our background
    and rendering task queues).

With this change committed, implementing resource loading
in WebAssembly should be straightforward.