VIRO-2444: yellow boxes on characters bug
This was caused because we weren't setting the
internal format explicitly in the texture constructor
used for font glyphs. While that value almost always
was initialized to RGBA8, it was unpredictable. On
WebAssembly it initialized to something different more
often, making the bug easier to find on that platform.
(it would initialize to YCbCR).