VIRO-3138 [iOS sim]: only use skinning.glsl when animating
As discovered previously, some drivers crash when a
UBO layout in an active shader has nothing bound to it.
This is the case with the iOS simulator and may be the case
with WebAssembly (explaining why particles and skinning
always crash that platform).

This commit fixes the issue with skinning, such that we only
include the skinning vertex shader, and its corresponding
UBO layout, if skinning is active for that material, meaning
its bone UBO will always be binding something.

To do this we had to ensure that #include directives can be
added to shader modifiers.