Retrieve Texture internalFormat from Image
Previously, there was some redundancy in the API
whereby a Texture needed an internalFormat parameter
even if it had an Image. This commit changes this so
that the Image itself sets its internalFormat (as it alone
really knows what the internalFormat is), and the Texture
just pulls the internalFormat from the Image.

This cleans up a number of areas in the codebase.