VIRO-3127 [font spec]: Android implementation
Android implementation of a font specification system.
Viro will now read the fonts.xml file on the device it's
running on, and parse it into various available font families.
When a request comes in for a typeface of a given
name, weight, and style, Viro will find the closest weight
of the font that matches the style. If the font is not found,
the default font will be returned with the desired weight
and style.