VIRO-3196: fallback fonts on iOS were not styled
We have to recreate the font with the attributes in order
to get the correct styling on them, in cases where
CoreText does not give us the requested font.

Documented the font table extraction methods.

SF font was causing issues (Freetype would not recognize
its font weight) so change the iOS default to Helvetica
Neue.

Finally, automatically select the system font if an empty
string is given for the family name.