Return ARPlane Polygon Vertex Data relative to Center
Instead of returning vertices that are relative to the ARAnchor of a given ARPlane,
we follow the same approach as surfaces that are placed in the center of ar planes
in the React bridge. We also align the behavior to follow Android: where the
polygon vertices are given in reference to the center of the ARPlane.