Fix Improper Placement of ARPlaneAnchors for ARCore
-  VROARPlane in iOS has a root transform, and as well as a center that is placed relative to
that transform. As the plane “grows” the center position will shift to accommodate the plane.
However, ARCore does not have similar behavior, and instead clamps the center of the ARPlane
at 0,0,0 (bug). The fix is to remove the enforcing of (0,0,0) and have the center location
be placed in the center of the plane as usual. This will align Android's ARPlane behavior
along with iOS.

- Ensure that the center position of a plane in ARCore also aligns with the same behavior
as ARKit, where the Y component is zero-ed out and ignored for the local center position,
and updated for the ARPlane's root position.