VIRO-3194: per-character typeface selection
Refactor text system to support typeface *collections*.
When creating a VROText, users now indicate the collection
of typefaces they would like to use to render the text. This
enables mixed-language text, whereby a user can provide
an English typeface and a Chinese typeface, for example,
to render mixed English / Chinese text.

The VROTypefaceCollection can iterate through a unicode
string and determine, for each character, what typeface is
best for rendering that character.

The best typeface is chosen by checking against each
typeface's "coverage", which is a sparse bit set indicating
what characters a typeface can render. The coverage is
generated from the raw 'cmap' TTF/OTF table.

Limited support for variant sequences is introduced here,
in that a typeface that supports a variant sequence will be
chosen over one which does not. Note that the font
*rendering* system still does not support variant sequences,
but some stubs were added to this effect.

The first use of this new feature will be to load the CJK TTC
file as a single typeface collection. In the future, we will load
all system fonts as a single collection, so that any combination
of text -- from any language -- can be rendered by default.