Driver should create typeface collections
The driver now supports creating typeface collections,
given a comma-separated list of typeface names.

This means collections of typefaces can be directly
passed in through the React bridge as well. In addition,
now that the driver creates collections instead of sole
typefaces, we can add logic on Android to load all
typefaces in a TTC file when given the font name for
a TTC.