Restore previous Text API and deprecate it
A previous commit changed the public Text API.
Instead of doing this, restore the former Text API
so there are no breakages in client code, but do
deprecate this API to encourage users to use the
builder class.

Make the new constructor with font weight and
style private, used only by the builder.