VIRO-3243: typeface crash on exit
We were destroying the freetype library
prior to calling FT_Done on some freetype
faces, leading to the crash. To fix this, detect
if the FT library has already been destroyed.

While working on this also discovered some
threading violations that forced an internal
API change. Previously, the typeface collections
were being created on the UI thread in Java: all
of this must happen on the rendering thread,
since FT is not threadsafe.

As a result we now pass all the parameters
needed to create the typeface into ViroText,
and ViroText itself will create the typeface. This
is fine since typefaces are cached and shared
anyway.