VIRO-3215: key validator exp falloff overflow
The time to wait would overflow resulting in a
negative number, causing a crash. This would
only occur at 54 retries. This does not happen
on iOS because we cap the number of retries.

To fix this for now we cap the falloff exponent.
Will regroup to consider capping the retries on
Android as well to match behavior on iOS.