VIRO-2934: use premultiplied alpha in Gaussian blur
Bloom was causing widespread distortion in AR scenes;
this was because we were adding the alpha values
from the bloom pass and the main HDR pass, which
resulted in undefined behavior.

A mitigation for this would have been capping the alpha
at 1.0 or just using the alpha value from the HDR buffer,
but that prevents blur from working on scenes with
transparent backgrounds (because the 'blur' could then
not appear over any area that had alpha=0 in the base
HDR buffer).

Instead fix this properly by using premultiplied alpha
blending in the Gaussian blur pass, and ensuring that
we set the correct blend mode when compositing the
blend passes together.