Fix bug in the ARRetail app where clicking anywhere moves the model
The shadow Surface was sized 20x20 meters, and that caused an issue because
when the user taps anywhere on the screen the surface receives an event which
bubbles up to the mProductModelGroup (because the Surface didn't handle the
event). The mProductModelGroup then handles the touch event because it has
an onDragListener set which then causes the model to move. The solution is
to call setIgnoreEventHandling() w/ true on the shadowNode to prevent it
from receiving/triggering events.