Do not overwrite anchor detection settings
Android was overwriting plane detection settings,
making it Horizontal. This could occur even after
a different plane detection was set in ViroReact.

Fixed this by making *Horizontal* plane detection
the default, specified in VROARScene, which is
"ground truth" for this setting. ViroReact can later
this setting through the anchorDetectionTypes
property. ViroCore can change this setting via the
setPlaneFindingMode in VROSceneRendererARCore.