[wasm] renderer test packaging and support
Part 1 of supporting the renderer tests. Instead of
having them as one monolithic test suite, as we
do on iOS and Android, break the tests up into
separate HTML files. This way we don't have to
preload all the resources for every test just to run
one of them.

This commit also demonstrates how to use Viro
as a library in WASM, in that we create an
additional CMakeLists file for the tests that calls
in to the master CMakeLists to build the renderer
library.

Currently all linking is done on the final executable,
nothing is "linked" to the renderer library.

Currently this only runs the FBX test.