[wasm] IDE support through CLion
Add IDE support to the WASM builds with CLion.
CLion builds an IntelliJ project off a CMakeLists file.
Added a new README.md that provides
instructions on how to do both command line and
CLion builds for WASM.

Most of the work here was moving script-level
functionality (like copying the preload directory)
over to CMake itself, so that the IDE could utilize
those same operations. The other notable change
was moving to a target-based system for choosing
what test to run, which is a much better fit.

Running with CLion works as well, as explained in the
Readme.