VIRO-3347: fix assets with spaces on Android
Assets with spaces require escaping on iOS, where
when loading from an NSBundle we use an actual URL,
whereas on Android when using file:///android_asset
we don't *actually* use a URL, we use the AssetManager,
so the escaping process actually breaks the search by
inserting %20.

Fix is to check if the URL is android_asset and, if so,
do not escape it.