VIRO-3408: shimmering on PBR models
This was caused because the FBX loader wasn't setting
the correct filtering settings on PBR textures, so mipmapping
was disabled.

Fix by forcing PBR models to use the defaults (linear). There
is no way to actually set this in the FBX exporter so it's ok
to force it here. Also made the change on the FBX exporter
to set linear on all textures.