VIRO-3383: deferred depth writes for hierarchical rendering
Far away flex-views (and any hierarchically rendered objects
in general) were suffering from Z-fighting. To fix this, change
how we render hierarchical objects:

1. Sort by hierarchy ID and depth, above transparency. This
    ensures that we render one hierarchy at a time.

2. During rendering, if we encounter a hierarchy, disable depth
    writes.

3. When finished rendering a hierarchy, render the parent of
    the hierarchy to the depth buffer only.

The effect of this is we use painter's algorithm for rendering the
hierarchy itself, but the hierarchy uses the depth buffer with
respect to objects outside the hierarchy.