VIRO-3385: Local Sound Test does not play
- When playing sounds in react, a full sound path in the format of file://path_to/sound.mp4
is already provided to VROAudioPlayeriOS. However, within the player, we also then append
yet an additional file// string to the already valid path with the fileURLWithPath API,
effectively making it an invalid path. The fix is to check for this prefix, and use the
URLWithString API if the path is already fully valid.

- Added additional soundDidFail checks that notifies developers if the player has failed
to load.