VIRO-3320 passert_thread issue with ARImageMarker
Some users are getting a pabort error when using ARImageMarker, but
the error location is within the VROThreadRestricted class, thus
obscuring the real error location. Making this a warning, so we still
see the logs, but don't immediately crash. This is to mitigate any
chance that we are too aggressively thread restricting.