VIRO-3432 [Android] Crash when rotating screen with ARImageTarget
We were re-initializing the ARTrackingSession when we get a
setDisplayGeometry call which was not being posted on the render
thread. This resulted in the ARTrackingSession "stopping" the
VROTextureReader on a different thread which results in it removing
itself from the frame synchronizer. In the meantime, the frame
synchronizer was trying to loop through its list of listeners on the
render thread and due to threading, it was occasionally accessing
a pointer that doesn't exist.