Native test module for running renderer tests
On Android we have the renderertest module, but
that was used for both Java-based rendering tests
(e.g. to debug github issues) and the native renderer
test harness. To switch between these we had to
flip a native constant flag, which was error-prone
and could be mistakenly checked in.

Introduce a new module, nativetest, that is specifically
intended to create the native renderer test harness
for running those tests.

This has the added benefit of reducing the asset
size of ViroActivity tests, so that these launch much
faster (most of the assets in renderertest were used
by the native tests).