VIRO-3393: oversaturation of Mali GPUs
Mali GPUs were not correctly reporting when they
had acquired an sRGB framebuffer. That is,
eglQuerySurface was reporting that they were
*not* acquiring an sRGB framebuffer even when
they were. As a result, we would turn on linear
software correction, so we would be *double*
gamma correcting the scene.

This fix adds a new GPU type, Mali. When we
detect a Mali GPU, we will simply assume we have
an sRGB framebuffer and thereby forcefully disable
software gamma correction. There is no known
method to detect if sRGB framebuffers are
returned by Mali so this will suffice for now.