Per-material tone-mapping implementation
Implement per-material tone mapping using MRT.
Previously we were going to do this with stencils,
but that required copying the stencil buffer over and
over, followed by two full-screen blit operations, which
was too slow. This new method is simpler: we write
out a tone-mapping mask to a secondary buffer,
and then read from that texture when performing the
actual tone mapping.

Some other changes were necessary to implement
this:

1. Introduced a new render target, because we can
    no longer write to the _hdrTarget when doing post
    processing effects (because if we did so we would
    overwrite the tone mapping mask, which we need
    for tone mapping)

2. Simplified the post processing effects API, so it
    returns the target with the final result. This eliminates
    a potential extra blit we were performing.