Allow silent failure of shader linkage to fix Adreno 530
The Adreno 530 fails to link shader programs that use
multiple render targets in a ViroSceneView (e.g. in a
TextureView).

This linking failure, however, is not deterministic; it almost
always succeeds after trying a second time. To work
around this driver bug, then, this commit introduces silent
linking failures: instead of hard crashing when a shader
fails to link, we instead just don't render that object
geometry for that frame, and try to compile and link the
shader again the next frame.