VIRO-3318: expose ARCore ambient light color
Expose ARCore's color correction, which is essentially
its estimated ambient light color. This matches the
"color temperature" value of ARKit. However, to unify
both under the same API, we migrate ARKit to return
an RGB color value as well.

This causes a breaking API change in that color temperature
is changed to RGB color for all AR light estimations.