VIRO-3626: create universal framework for react-viro
For simulator support in react-viro, create a universal
ViroKit.framework that supports both x86_64 and
arm64. To do this, modify the 'Copy Frameworks to Bridge'
Build Phase such that:

1. If we're building for the Simulator, copy the framework
    to react-viro/ios/dist/ViroRenderer/x86_64

2. If we're building for device, copy the framework to
    react_viro/ios/dist/ViroRenderer/armv7_arm64

3. Regardless of what we're building for, use the lipo command
    to combine the two together (this assumes both exist)
    and place the resultant universal framework in
    react_viro/ios/dist/ViroRenderer. Note this last path is where
    the single framework used to be, prior to this commit.

No changes were required in react-viro. Furthermore,
Cocoapods will automatically strip out the architectures in
ViroKit.framework that are not used (this operation takes
place in a Build Phase for react-viro; it should also exist
for all other ViroReact projects)