VIRO-3593: separable atomic recomputation for Node
This commit enables the ability of Nodes to recompute
all of their atomic variables from any thread. Meaning,
we can set the position of a Node on the UI thread, and
its atomic variables for computedTransform and other
variables will be immediately recomputed.

Previously, we had to wait until the next render cycle
was complete, and syncAtomicRenderProperties was
called, for these variables to be updated. This caused
problems, however, where the user's expectation of
what was set was not met.

This behavior manifested in the bug listed above, where
convertLocalCoordinatsToWorld would return stale
values for any recently moved Node.