VIRO-2162: run tests on UI thread and cleanup
This makes all tests run on the UI thread. The new
runUITest() function takes a lambda which will run the
test on the UI thread, and then wait on the Test
thread for either button to be tapped.

Additionally, after a test is run, the resetTestState()
method is called (also on the UI thread). By default,
this clears the Scene and recreates it by invoking
configureTestScene() again. However, individual
tests can override this if it's inefficient. This way the
state is properly reset after each test.