VIRO-3592: Incorrect Dragging Behavior for Parented child nodes.
The dragging behavior of child subnodes was not properly dragging them to
correct transformations relative to it's parent, resulting in weird dragging
behavior, especially if the parent node itself has additional rotation or
translation. The fix is to always perform drag operations in world
transformations (they should be anyways, since the hit positions themselves
that are used to calculate dragTo locations are in world coordinates) and
then to conver these back into local transforms to be set into the Node.