VIRO-3698: use entire VROData instead of computing size
GLTF uses the byte offset in VROGeometrySource to compute
not just the offset of an attribute within a vertex (e.g. the offset
within the stride), but also to compute the *start* of a buffer
within the data array. This works fine with OpenGL but it did
perturb our dataSize calculations, which we used to limit the
size of the data being bound to OpenGL.

It turns out this size isn't even necessary since we're using
VROData for our data buffers, and these have a fixed length
that we can utilize for the size.