VIRO-3706: refactor ARCore implementation to use anchors
The existing ARCore implementation relies on trackables,
updating them every frame and synchronizing those with
Viro objects. We need to adjust this so that we get the
tranforms from anchors *acquired from* trackables, instead
of trackables themselves. This is necessary in order for
cloud anchors to work, since we can't host trackables in the
cloud.

ARCore consists of two concepts: trackable and anchor.
Trackables are detected real-world objects, like horizontal
and vertical planes, or image targets. Anchors are virtual
objects that are attached to the real world, either relative to a
trackable or relative to an AR hit result.

Unlike ARCore, Viro (and ARKit) merge these concepts
together: trackables *are* anchors. In order to bridge this
conceptual difference with ARCore, this commit changes
our implementation so that we create one ARCore anchor
for every ARCore trackable found. That anchor is attached to
trackable in VROARAnchorARCore.

We then use that created anchor for deriving transformations
for Viro objects. We continue to use the underlying trackable
of that anchor to update other properties in our Viro anchors
(for example, plane extent).