VIRO-3706: use anchors for hit-tests
The existing implementation of hit-test was not
correct for ARCore: we were just returning the
position of hit locations, and users would set their
objects to those positions. However, with ARCore
(unlike ARKit) the world coordinate system from
frame to frame is *not* consistent. You *must* use
anchors to position objects.

To make this work, this commit introduces "manual"
anchors: anchors that you can create from a hit
result. Users can now call hitResult.createAnchoredNode()
to create an ARNode with an attached anchor, to
which they can add objects. When they're finished
with the node, they call ARNode.detach().

This is a large change. Highlights are:

1. There are now managed anchors and manual anchors.
    Managed anchors are those that we automatically create
    for each trackable we find. Manual anchors are those
    that are created by users through createAnchoredNode.

2. The VROARHitResult is now a long-lived object, so it had to be
     turned into a shared_ptr everywhere.

3. Strong reference cycle has been introduced between
    VROARNode and VROARAnchor in order to make manual
    anchors and nodes stay alive. This reference cycle is correctly
    broken whenever an anchor is detached.

4. All adding and removing of anchors, both manual and managed,
    occurs through VROARSession::addAnchor and
    VROARSession::removeAnchor.